separate logging from scheduling
