+++ src/commands/test/test_command.py
@@ -4,9 +4,10 @@ from pathlib import Path
 from typing import TYPE_CHECKING, List, Optional, Pattern
 
 from src.cli.command import Command
-from src.commands.test.reporter import ReporterCoordinator
 from src.commands.test.runner import TestRunner
 from src.commands.test.test_collector import TestCollector
+from src.commands.test.test_live_logger import TestLiveLogger
+from src.commands.test.test_scheduler import TestScheduler
 from src.utils.protostar_directory import ProtostarDirectory
 
 if TYPE_CHECKING:
@@ -96,6 +97,6 @@ class TestCommand(Command):
         )
         test_collector_result.log(logger)
 
-        ReporterCoordinator(logger, worker=TestRunner.worker).run(
+        TestScheduler(TestLiveLogger(logger), worker=TestRunner.worker).run(
             include_paths=include_paths, test_collector_result=test_collector_result
         )
